+++ src/main/java/com/ferreusveritas/dynamictrees/systems/dropcreators/DropCreatorSeed.java
@@ -73,7 +73,7 @@ public class DropCreatorSeed implements IDropCreator {
 	}
 
 	@Override
-	public List<ItemStack> getLogsDrop(World world, Species species, BlockPos breakPos, Random random, List<ItemStack> dropList, int volume) {
+	public List<ItemStack> getLogsDrop(World world, Species species, BlockPos breakPos, Random random, List<ItemStack> dropList, float volume) {
 		return dropList;
 	}
