+++ src/main/java/com/ferreusveritas/dynamictrees/systems/dropcreators/DropCreatorStorage.java
@@ -93,7 +93,7 @@ public class DropCreatorStorage implements IDropCreatorStorage {
 		return dropList;
 	}
 	
-	public List<ItemStack> getLogsDrop(World world, Species species, BlockPos rootPos, Random random, List<ItemStack> dropList, int volume) {
+	public List<ItemStack> getLogsDrop(World world, Species species, BlockPos rootPos, Random random, List<ItemStack> dropList, float volume) {
 		dropList = makeDropListIfNull(dropList);
 		
 		for(IDropCreator dropCreator : dropCreators.values()) {
